replaced prev definition in all places, removed react syntex highlighter
