+++ package.json
@@ -94,7 +94,6 @@
     "react-redux": "8.0.1",
     "react-router-dom": "6.3.0",
     "react-select": "^5.4.0",
-    "react-syntax-highlighter": "^15.5.0",
     "react-transition-group": "4.4.2",
     "react-use": "17.4.0",
     "react-use-websocket": "^4.2.0",
