+++ src/components/organisms/EntityDetails/EntityDetailsContent/Settings/SettingsDefinition/utils.ts
@@ -1,10 +1,13 @@
-import {UseQuery} from '@reduxjs/toolkit/dist/query/react/buildHooks';
+import {MutationDefinition} from '@reduxjs/toolkit/dist/query';
+import {UseMutation, UseQuery} from '@reduxjs/toolkit/dist/query/react/buildHooks';
 import {QueryDefinition} from '@reduxjs/toolkit/src/query/endpointDefinitions';
 
 import {Entity} from '@models/entity';
 
-import {useGetTestSuiteDefinitionQuery} from '@services/testSuites';
-import {useGetTestDefinitionQuery} from '@services/tests';
+import {useGetTestSuiteDefinitionQuery, useUpdateTestSuiteDefinitionMutation} from '@services/testSuites';
+import {useGetTestDefinitionQuery, useUpdateTestDefinitionMutation} from '@services/tests';
+
+import {YamlEditBody} from '@src/models/fetch';
 
 import {externalLinks} from '@utils/externalLinks';
 
@@ -14,16 +17,19 @@ export const settingsDefinitionData: Record<
     description: string;
     helpLinkUrl: string;
     query: UseQuery<QueryDefinition<string, any, any, string>>;
+    mutation: UseMutation<MutationDefinition<YamlEditBody, any, any, any, any>>;
   }
 > = {
   'test-suites': {
     description: 'Validate and export your test suite configuration',
     helpLinkUrl: externalLinks.createTestSuite,
     query: useGetTestSuiteDefinitionQuery,
+    mutation: useUpdateTestSuiteDefinitionMutation,
   },
   tests: {
     description: 'Validate and export your test configuration',
     helpLinkUrl: externalLinks.createTest,
     query: useGetTestDefinitionQuery,
+    mutation: useUpdateTestDefinitionMutation,
   },
 };
