+++ src/components/pages/Executors/ExecutorDetails/ExecutorSettings/ExecutorDefinition/ExecutorDefinition.tsx
@@ -1,45 +1,26 @@
-import {useContext} from 'react';
+import {Definition} from '@molecules';
 
-import {Form} from 'antd';
+import {useAppDispatch, useAppSelector} from '@redux/hooks';
+import {selectCurrentExecutor, setCurrentExecutor} from '@redux/reducers/executorsSlice';
 
-import {CopyButton, DownloadButton, Pre} from '@atoms';
+import {useGetExecutorDefinitionQuery, useUpdateExecutorDefinitionMutation} from '@services/executors';
 
-import {MainContext} from '@contexts';
-
-import useLocation from '@hooks/useLocation';
-import useSecureContext from '@hooks/useSecureContext';
-
-import {ConfigurationCard, Definition as DefinitionContent} from '@molecules';
-
-import {useAppSelector} from '@redux/hooks';
-import {selectCurrentExecutor} from '@redux/reducers/executorsSlice';
-
-import {useGetExecutorDefinitionQuery} from '@services/executors';
+import {testkubeCRDBases} from '@utils/externalLinks';
 
 const ExecutorDefinition = () => {
-  const {isClusterAvailable} = useContext(MainContext);
+  const executor = useAppSelector(selectCurrentExecutor)!;
 
-  const executor = useAppSelector(selectCurrentExecutor);
-  const {data: definition = '', isLoading} = useGetExecutorDefinitionQuery(executor?.name, {skip: !isClusterAvailable});
-  const isSecureContext = useSecureContext();
-  const filename = useLocation().lastPathSegment;
+  const dispatch = useAppDispatch();
 
   return (
-    <Form name="definition-form">
-      <ConfigurationCard title="Definition" description="Validate and export your container executor configuration">
-        {definition ? (
-          <DefinitionContent content={definition}>
-            {isSecureContext ? (
-              <CopyButton content={definition} />
-            ) : (
-              <DownloadButton filename={filename} extension="yaml" content={definition} />
-            )}
-          </DefinitionContent>
-        ) : (
-          <Pre>{isLoading ? ' Loading...' : ' No definition data'}</Pre>
-        )}
-      </ConfigurationCard>
-    </Form>
+    <Definition
+      useGetDefinitionQuery={useGetExecutorDefinitionQuery}
+      useUpdateDefinitionMutation={useUpdateExecutorDefinitionMutation}
+      label="executor"
+      setEntity={dispatch(setCurrentExecutor)}
+      name={executor.name}
+      crdUrl={testkubeCRDBases.executors}
+    />
   );
 };
