+++ src/services/tests.ts
@@ -1,7 +1,7 @@
 import {createApi} from '@reduxjs/toolkit/query/react';
 
 import {Artifact} from '@models/artifact';
-import {MetadataResponse} from '@models/fetch';
+import {MetadataResponse, YamlEditBody} from '@models/fetch';
 import {Test, TestFilters, TestWithExecution} from '@models/test';
 
 import {dynamicBaseQuery, memoizeQuery, paramsSerializer} from '@utils/fetchUtils';
@@ -108,6 +108,16 @@ export const testsApi = createApi({
         method: 'POST',
       }),
     }),
+    updateTestDefinition: builder.mutation<any, YamlEditBody>({
+      query: body => {
+        return {
+          url: `/tests/${body.name}`,
+          method: 'PATCH',
+          headers: {'Content-Type': 'text/yaml'},
+          body: body.value,
+        };
+      },
+    }),
   }),
 });
 
@@ -141,4 +151,5 @@ export const {
   useRunTestMutation,
   useAbortTestExecutionMutation,
   useAbortAllTestExecutionsMutation,
+  useUpdateTestDefinitionMutation,
 } = testsApi;
